refactor(auto prom. links): move to renderers
