+++ baker/formatWordpressPost.tsx
@@ -5,7 +5,7 @@ import * as lodash from "lodash"
 import * as React from "react"
 import * as ReactDOMServer from "react-dom/server"
 import { HTTPS_ONLY } from "../settings/serverSettings"
-import { getMediaThumbnailUrl, getPostBySlug, getTables } from "../db/wpdb"
+import { getTables } from "../db/wpdb"
 import Tablepress from "../site/Tablepress"
 import { GrapherExports } from "../baker/GrapherBakingUtils"
 import * as path from "path"
@@ -22,10 +22,7 @@ import {
 import { bakeGlobalEntitySelector } from "./bakeGlobalEntitySelector"
 import { Footnote } from "../site/Footnote"
 import { LoadingIndicator } from "../grapher/loadingIndicator/LoadingIndicator"
-import {
-    PROMINENT_LINK_CLASSNAME,
-    renderAutomaticProminentLink,
-} from "../site/blocks/ProminentLink"
+import { PROMINENT_LINK_CLASSNAME } from "../site/blocks/ProminentLink"
 import { countryProfileSpecs } from "../site/countryProfileProjects"
 import { formatGlossaryTerms } from "../site/formatGlossary"
 import { getMutableGlossary, glossary } from "../site/glossary"
@@ -37,7 +34,6 @@ import {
     formatDataValue,
     formatLinks,
     getHtmlContentWithStyles,
-    isStandaloneInternalLink,
     parseKeyValueArgs,
 } from "./formatting"
 import { mathjax } from "mathjax-full/js/mathjax"
@@ -54,9 +50,7 @@ import {
     getLegacyVariableDisplayConfig,
 } from "../db/model/Variable"
 import { AnnotatingDataValue } from "../site/AnnotatingDataValue"
-import { Url } from "../clientUtils/urls/Url"
-import { logErrorAndMaybeSendToSlack } from "./slackLog"
-import { formatWordpressEditLink } from "../site/LongFormPage"
+import { renderAutomaticProminentLinks } from "./siteRenderers"
 
 const initMathJax = () => {
     const adaptor = liteAdaptor()
@@ -300,57 +294,7 @@ export const formatWordpressPost = async (
 
     // SSR rendering of Gutenberg blocks, before hydration on client
     renderBlocks(cheerioEl)
-
-    // render automatic prominent links
-    const anchorElements = cheerioEl("a").toArray()
-    await Promise.all(
-        anchorElements.map(async (anchor) => {
-            if (!isStandaloneInternalLink(anchor, cheerioEl)) return
-            const url = Url.fromURL(anchor.attribs.href)
-            if (!url.slug) return
-
-            let targetPost
-            try {
-                targetPost = await getPostBySlug(url.slug)
-            } catch (err) {
-                // not throwing here as this is not considered a critical error.
-                // Standalone links will just show up as such (and get
-                // netlify-redirected upon click if applicable).
-                logErrorAndMaybeSendToSlack(
-                    new Error(
-                        `${post.title} (${formatWordpressEditLink(
-                            post.id
-                        )}) wants to convert ${
-                            anchor.attribs.href
-                        } into a prominent link but this URL doesn't match any post. This might be because the URL of the target has been recently changed.`
-                    )
-                )
-            }
-
-            if (!targetPost) return
-
-            let mediaThumbnailUrl
-
-            if (targetPost.imageId) {
-                mediaThumbnailUrl = await getMediaThumbnailUrl(
-                    targetPost.imageId
-                )
-                if (mediaThumbnailUrl) {
-                    mediaThumbnailUrl = formatLinks(mediaThumbnailUrl)
-                }
-            }
-
-            const rendered = renderAutomaticProminentLink(
-                anchor.attribs.href,
-                targetPost.title,
-                mediaThumbnailUrl
-            )
-
-            const $anchorParent = cheerioEl(anchor.parent)
-            $anchorParent.after(rendered)
-            $anchorParent.remove()
-        })
-    )
+    await renderAutomaticProminentLinks(cheerioEl, post)
 
     // Extract blog info content
     let info = null
